+++ internal/oci/testdata/authenticated.yaml
