+++ internal/oci/testdata/happy-path.yaml
