+++ internal/oci/testdata/multiple-artifacts.yaml
