+++ internal/oci/testdata/no-artifacts.yaml
