+++ internal/oci/testdata/wrong-authentication.yaml
